converting data in memory to float for speed

+++ utils/memory.py
@@ -26,18 +26,18 @@ class Memory():
 
     def sample(self, batch_size):
         transitions = self.transition(*zip(*self.rng.sample(self.buffer, batch_size)))
-        state = torch.Tensor(transitions.state).to(self.device)
-        action = torch.stack(transitions.action).to(self.device)
-        reward = torch.Tensor(transitions.reward).to(self.device)
-        next_state = torch.Tensor(transitions.next_state).to(self.device)
-        mask = torch.Tensor(transitions.mask).to(self.device)
+        state = torch.Tensor(transitions.state).float().to(self.device)
+        action = torch.stack(transitions.action).float().to(self.device)
+        reward = torch.Tensor(transitions.reward).float().to(self.device)
+        next_state = torch.Tensor(transitions.next_state).float().to(self.device)
+        mask = torch.Tensor(transitions.mask).float().to(self.device)
         return self.transition(state, action, reward, next_state, mask)
     
     def replay(self):
         transitions = self.transition(*zip(*self.buffer))
-        state = torch.Tensor(transitions.state).to(self.device)
-        action = torch.stack(transitions.action).to(self.device)
-        reward = torch.Tensor(transitions.reward).to(self.device)
-        next_state = torch.Tensor(transitions.next_state).to(self.device)
-        mask = torch.Tensor(transitions.mask).to(self.device)
+        state = torch.Tensor(transitions.state).float().to(self.device)
+        action = torch.stack(transitions.action).float().to(self.device)
+        reward = torch.Tensor(transitions.reward).float().to(self.device)
+        next_state = torch.Tensor(transitions.next_state).float().to(self.device)
+        mask = torch.Tensor(transitions.mask).float().to(self.device)
         return self.transition(state, action, reward, next_state, mask)
